Class type Csv.in_obj_channel

The most basic input object for best interoperability.

method input : Stdlib.Bytes.t -> int -> int -> int

input buf ofs len reads up to len octets from the channel and puts them in the substring buf.[ofs .. ofs+len-1]. Returns the number of octets actually read (and stored). When the channel is non-blocking, and there are currently no bytes to read, the number 0 will be returned.

raises End_of_file

when there are no more bytes to read.

method close_in : unit -> unit

Closes the channel for input.