Class type Csv.out_obj_channel

The most basic output object for best interoperability.

method output : Stdlib.Bytes.t -> int -> int -> int

output s ofs len writes up to len bytes of the substring s.[ofs .. ofs+len-1]. Return the number of bytes actually written. When the channel is non-blocking, and there are currently no bytes to write, the number 0 must be returned.

method close_out : unit -> unit

Flushes the buffer, if any, and closes the channel for output.